+++ util/common/src/main/java/org/hibernate/search/util/common/reflect/spi/AbstractTypeOrdering.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.mapper.pojo.util.spi;
+package org.hibernate.search.util.common.reflect.spi;
 
 import java.util.LinkedHashSet;
 import java.util.Set;
@@ -12,9 +12,9 @@ import java.util.stream.Stream;
 
 
 
-public abstract class AbstractPojoTypeOrdering<T> {
+public abstract class AbstractTypeOrdering<T> {
 
-	protected AbstractPojoTypeOrdering() {
+	protected AbstractTypeOrdering() {
 	}
 
 	public Stream<? extends T> getAscendingSuperTypes(T subType) {
